Revert "Refractor and restructure the Server codebase and add new live match data feature"
