--- ServerStuff/.gitignore
@@ -1,10 +1,0 @@
-# Ignore Python cache files
-**/__pycache__/*
-# Ignore all files in the Holodelta web
-**/Holodelta_web/*
-# Ignore Python bytecode files
-**/*.pyc
-# Ignore environment variable files
-**/.env*
-# SQLite database files
-**/*.db
