+++ ServerStuff/banlists/current.json
