--- ServerStuff/requirements.txt
@@ -1,6 +1,0 @@
-fastapi
-uvicorn[standard]
-gunicorn
-sqlalchemy
-databases[sqlite]
-python-dotenv
